refactor: setTargetBlankの中に詰め込んでいたロジックの別ファイル切り出し
