+++ src/features/target-blank/contentScripts/setTargetBlank.ts
@@ -1,17 +1,8 @@
-import { TargetBlankSettings } from "../types"
-import { convertDomElement, DOMElement, getPreviewObserver, SELECTOR } from "./domAdapter"
+import { DOMElement, TargetBlankSettings } from "../types"
+import { convertDomElement, SELECTOR } from "./domAdapter"
+import { effectsDom } from "./effectsDom";
 
 export const setTargetBlank = (settings: TargetBlankSettings) => {
-    const previewBox = document.querySelector(SELECTOR.previewWrapper);
-
-    if (previewBox) {
-        const observer = getPreviewObserver({ settings, effectToDom })
-        observer.observe(previewBox, {
-            childList: true,
-            subtree: true
-        })
-    }
-
     const rawElements = document.querySelectorAll(SELECTOR.main)
     let elements: DOMElement[] = []
 
@@ -19,32 +10,5 @@ export const setTargetBlank = (settings: TargetBlankSettings) => {
         elements = convertDomElement(rawElements)
     }
 
-    effectToDom(elements, settings)
-}
-
-const effectToDom = (elements: DOMElement[], settings: TargetBlankSettings) => {
-    const excludeUrlList = settings.excludeUrlList.map(excludeUrl => excludeUrl.host)
-
-    elements.forEach(element => {
-        if (!settings.alwaysOpenOtherTab) {
-            return
-        }
-        if (settings.inKibelaLinkOpenSameTab && (element.isUserMention || element.isSameSiteLink)) {
-            return
-        }
-        if (element.isAnchorLink) {
-            return
-        }
-        try {
-            const href = element.isUserMention || element.isSameSiteLink ? `${location.origin}${element.href}` : element.href
-            const { host } = new URL(href)
-            if (excludeUrlList.includes(host)) {
-                return
-            }
-        } catch (e) {
-            console.error("Caused URL:", element.href);
-            console.error(e);
-        }
-        element.setTargetBlankAttribute(element.rawElement)
-    })
+    effectsDom({ elements, settings })
 }
